feat: split settings into two-level navigation (closes #16, #29)
Replaces the 2500-line settings.vue monolith with a sidebar wrapper
and six independently navigable sub-pages.

- settings.vue → sidebar on desktop, <NuxtPage /> slot
- settings/index.vue → mobile section list; desktop auto-redirects to /settings/display
- settings/display.vue → theme, color mode, motion, time format, matrix, habits/today, log mode
- settings/features.vue → feature flags, Pomodoro config, health setup modal
- settings/notifications.vue → permission, test buttons, notification log + scheduled list
- settings/permissions.vue → install, exact alarms, battery, persistent storage, media perms
- settings/data.vue → export JSON/SQLite/Jots, import, clear, nuke OPFS (all modals co-located)
- settings/more.vue → About, Diagnostics, Here be dragons (collapsible)

Philosophy note shown at bottom of mobile index page and in the desktop sidebar.
Deep links work for all sub-pages; back button on mobile (sm:hidden on desktop).
