+++ app/pages/settings.vue
@@ -1,2515 +1,45 @@
 <script setup lang="ts">
-const db = useDatabase()
-const runtimeConfig = useRuntimeConfig()
-const { settings: appSettings, set: setAppSetting } = useAppSettings()
+const route = useRoute()
 
-// ─── JSON export ──────────────────────────────────────────────────────────────
-
-type ExportKey = keyof ExportSelection
-
-interface ExportGroup {
-  label: string
-  items: Array<{ key: ExportKey; label: string; parent?: ExportKey }>
-}
-
-const EXPORT_GROUPS: ExportGroup[] = [
-  {
-    label: 'Habits',
-    items: [
-      { key: 'habits', label: 'Habits' },
-      { key: 'completions', label: 'Completions', parent: 'habits' },
-      { key: 'habit_logs', label: 'Habit logs', parent: 'habits' },
-      { key: 'habit_schedules', label: 'Habit schedules', parent: 'habits' },
-      { key: 'reminders', label: 'Habit reminders', parent: 'habits' },
-    ],
-  },
-  {
-    label: 'Check-ins',
-    items: [
-      { key: 'checkin_templates', label: 'Templates' },
-      { key: 'checkin_questions', label: 'Questions', parent: 'checkin_templates' },
-      { key: 'checkin_responses', label: 'Responses', parent: 'checkin_questions' },
-      { key: 'checkin_reminders', label: 'Reminders', parent: 'checkin_templates' },
-    ],
-  },
-  {
-    label: 'Other',
-    items: [
-      { key: 'scribbles', label: 'Scribbles' },
-      { key: 'checkin_entries', label: 'Journal entries' },
-    ],
-  },
-  {
-    label: 'TODOs & Bored',
-    items: [
-      { key: 'todos', label: 'TODOs' },
-      { key: 'bored_categories', label: 'Bored categories' },
-      { key: 'bored_activities', label: 'Bored activities', parent: 'bored_categories' },
-    ],
-  },
+const SETTINGS_SECTIONS = [
+  { to: '/settings/display', label: 'Display', icon: 'i-heroicons-swatch' },
+  { to: '/settings/features', label: 'Features', icon: 'i-heroicons-squares-2x2' },
+  { to: '/settings/notifications', label: 'Notifications', icon: 'i-heroicons-bell' },
+  { to: '/settings/permissions', label: 'Permissions', icon: 'i-heroicons-shield-check' },
+  { to: '/settings/data', label: 'Data', icon: 'i-heroicons-circle-stack' },
+  { to: '/settings/more', label: 'More', icon: 'i-heroicons-ellipsis-horizontal-circle' },
 ]
-
-function defaultExportSelection(): ExportSelection {
-  return {
-    habits: true,
-    completions: true,
-    habit_logs: true,
-    habit_schedules: true,
-    reminders: true,
-    checkin_templates: true,
-    checkin_questions: true,
-    checkin_responses: true,
-    checkin_reminders: true,
-    scribbles: true,
-    checkin_entries: true,
-    todos: true,
-    bored_categories: true,
-    bored_activities: true,
-  }
-}
-
-const showExportModal = ref(false)
-const exporting = ref(false)
-const exportSel = reactive<ExportSelection>(defaultExportSelection())
-const exportErrors = ref<string[]>([])
-
-const exportAllKeys = Object.keys(defaultExportSelection()) as ExportKey[]
-const exportAllSelected = computed(() => exportAllKeys.every((k) => exportSel[k]))
-const exportNoneSelected = computed(() => exportAllKeys.every((k) => !exportSel[k]))
-
-function toggleExportAll() {
-  const next = !exportAllSelected.value
-  for (const k of exportAllKeys) exportSel[k] = next
-}
-
-function openExportModal() {
-  Object.assign(exportSel, defaultExportSelection())
-  exportErrors.value = []
-  showExportModal.value = true
-}
-
-function validateExportFk(): string[] {
-  type Rule = [child: ExportKey, parent: ExportKey, childLabel: string, parentLabel: string]
-  const rules: Rule[] = [
-    ['completions', 'habits', 'Completions', 'Habits'],
-    ['habit_logs', 'habits', 'Habit logs', 'Habits'],
-    ['habit_schedules', 'habits', 'Habit schedules', 'Habits'],
-    ['reminders', 'habits', 'Habit reminders', 'Habits'],
-    ['checkin_questions', 'checkin_templates', 'Questions', 'Check-in templates'],
-    ['checkin_responses', 'checkin_questions', 'Responses', 'Check-in questions'],
-    ['checkin_reminders', 'checkin_templates', 'Check-in reminders', 'Check-in templates'],
-    ['bored_activities', 'bored_categories', 'Bored activities', 'Bored categories'],
-  ]
-  return rules
-    .filter(([child, parent]) => exportSel[child] && !exportSel[parent])
-    .map(([, , childLabel, parentLabel]) => `${childLabel} require ${parentLabel}`)
-}
-
-async function downloadJson() {
-  if (!db.isAvailable) return
-  exportErrors.value = validateExportFk()
-  if (exportErrors.value.length > 0) return
-  exporting.value = true
-  try {
-    const data = await db.exportJsonData({ ...exportSel })
-    const payload = JSON.stringify(data, null, 2)
-    const url = URL.createObjectURL(new Blob([payload], { type: 'application/json' }))
-    try {
-      const a = document.createElement('a')
-      a.href = url
-      a.download = `habitat-${new Date().toISOString().slice(0, 10)}.json`
-      document.body.appendChild(a)
-      a.click()
-      document.body.removeChild(a)
-    } finally {
-      URL.revokeObjectURL(url)
-    }
-    showExportModal.value = false
-  } finally {
-    exporting.value = false
-  }
-}
-
-// ─── SQLite export ────────────────────────────────────────────────────────────
-
-const exportingDb = ref(false)
-
-async function exportSqlite() {
-  if (!db.isAvailable) return
-  exportingDb.value = true
-  try {
-    const bytes = await db.exportDb()
-    const url = URL.createObjectURL(new Blob([bytes], { type: 'application/x-sqlite3' }))
-    try {
-      const a = document.createElement('a')
-      a.href = url
-      a.download = `habitat-${new Date().toISOString().slice(0, 10)}.sqlite3`
-      document.body.appendChild(a)
-      a.click()
-      document.body.removeChild(a)
-    } finally {
-      URL.revokeObjectURL(url)
-    }
-  } finally {
-    exportingDb.value = false
-  }
-}
-
-// ─── JSON import ──────────────────────────────────────────────────────────────
-
-const importInput = ref<HTMLInputElement | null>(null)
-const showImportModal = ref(false)
-const importing = ref(false)
-const importPreview = ref<HabitatExport | null>(null)
-const importError = ref<string | null>(null)
-const importDone = ref(false)
-
-function openImport() {
-  importPreview.value = null
-  importError.value = null
-  importDone.value = false
-  importInput.value?.click()
-}
-
-async function onImportFileSelected(e: Event) {
-  const file = (e.target as HTMLInputElement).files?.[0]
-  if (!file) return
-  ;(e.target as HTMLInputElement).value = ''
-  try {
-    const raw = JSON.parse(await file.text()) as unknown
-    if (typeof raw !== 'object' || raw === null) {
-      importError.value = 'Invalid file.'
-      showImportModal.value = true
-      return
-    }
-    const ver = (raw as Record<string, unknown>).version
-    if (typeof ver !== 'number') {
-      importError.value = 'Missing version field — not a Habitat export.'
-      showImportModal.value = true
-      return
-    }
-    if (ver !== 1) {
-      importError.value = `Unsupported version ${ver}. This app supports version 1.`
-      showImportModal.value = true
-      return
-    }
-    importPreview.value = raw as HabitatExport
-    showImportModal.value = true
-  } catch {
-    importError.value = 'Could not parse file as JSON.'
-    showImportModal.value = true
-  }
-}
-
-function reloadPage() {
-  window.location.reload()
-}
-
-async function confirmImport() {
-  if (!importPreview.value || !db.isAvailable) return
-  importing.value = true
-  try {
-    await db.importJson(importPreview.value)
-    importDone.value = true
-    importPreview.value = null
-  } finally {
-    importing.value = false
-  }
-}
-
-// ─── Jots ZIP export ──────────────────────────────────────────────────────────
-
-interface VoiceNoteRaw {
-  id: string
-  blob: Blob
-  mimeType: string
-  created_at: string
-}
-
-interface ImageNoteRaw {
-  id: string
-  blob: Blob
-  mimeType: string
-  filename: string
-  created_at: string
-}
-
-function openJotsIdb(): Promise<IDBDatabase> {
-  return new Promise((resolve, reject) => {
-    const req = indexedDB.open('habitat', 2)
-    req.onupgradeneeded = (e) => {
-      const idb = req.result
-      if (e.oldVersion < 1) idb.createObjectStore('voice_notes', { keyPath: 'id' })
-      if (e.oldVersion < 2) idb.createObjectStore('image_notes', { keyPath: 'id' })
-    }
-    req.onsuccess = () => resolve(req.result)
-    req.onerror = () => reject(req.error)
-  })
-}
-
-function idbGetAllNotes(idb: IDBDatabase): Promise<VoiceNoteRaw[]> {
-  return new Promise((resolve, reject) => {
-    const req = idb.transaction('voice_notes', 'readonly').objectStore('voice_notes').getAll()
-    req.onsuccess = () => resolve(req.result as VoiceNoteRaw[])
-    req.onerror = () => reject(req.error)
-  })
-}
-
-function idbGetAllImages(idb: IDBDatabase): Promise<ImageNoteRaw[]> {
-  return new Promise((resolve, reject) => {
-    const req = idb.transaction('image_notes', 'readonly').objectStore('image_notes').getAll()
-    req.onsuccess = () => resolve(req.result as ImageNoteRaw[])
-    req.onerror = () => reject(req.error)
-  })
-}
-
-const showJotsExportModal = ref(false)
-const exportingJots = ref(false)
-const jotsExportSel = reactive({ text: true, voice: true, images: true })
-
-// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: sequential export pipeline
-async function exportJotsZip() {
-  exportingJots.value = true
-  try {
-    const files: Record<string, Uint8Array> = {}
-
-    if (jotsExportSel.text && db.isAvailable) {
-      const textJots = await db.getScribbles()
-      const json = JSON.stringify(
-        { version: 1, exported_at: new Date().toISOString(), text: textJots },
-        null,
-        2,
-      )
-      files['jots.json'] = new TextEncoder().encode(json)
-      for (const jot of textJots) {
-        const body = jot.title ? `${jot.title}\n\n${jot.content}` : jot.content
-        const ts = jot.updated_at.slice(0, 19).replace(/[:.]/g, '-')
-        files[`text/${ts}.txt`] = new TextEncoder().encode(body)
-      }
-    }
-
-    if (jotsExportSel.voice || jotsExportSel.images) {
-      const idb = await openJotsIdb()
-      if (jotsExportSel.voice) {
-        const notes = await idbGetAllNotes(idb)
-        for (const note of notes) {
-          const ext = note.mimeType.split('/')[1]?.split(';')[0] ?? 'audio'
-          const ts = note.created_at.slice(0, 19).replace(/[:.]/g, '-')
-          files[`voice/${ts}.${ext}`] = new Uint8Array(await note.blob.arrayBuffer())
-        }
-      }
-      if (jotsExportSel.images) {
-        const images = await idbGetAllImages(idb)
-        for (const img of images) {
-          const ext = img.mimeType.split('/')[1] ?? 'jpg'
-          const ts = img.created_at.slice(0, 19).replace(/[:.]/g, '-')
-          files[`images/${ts}.${ext}`] = new Uint8Array(await img.blob.arrayBuffer())
-        }
-      }
-    }
-
-    if (Object.keys(files).length === 0) return
-
-    const zipped = zipSync(files)
-    const url = URL.createObjectURL(
-      new Blob([zipped.buffer as ArrayBuffer], { type: 'application/zip' }),
-    )
-    try {
-      const a = document.createElement('a')
-      a.href = url
-      a.download = `habitat-jots-${new Date().toISOString().slice(0, 10)}.zip`
-      document.body.appendChild(a)
-      a.click()
-      document.body.removeChild(a)
-    } finally {
-      URL.revokeObjectURL(url)
-    }
-    showJotsExportModal.value = false
-  } finally {
-    exportingJots.value = false
-  }
-}
-
-// ─── Shared helpers ───────────────────────────────────────────────────────────
-
-/** Delete the entire IndexedDB 'habitat' database (voice notes). */
-function clearIdb(): Promise<void> {
-  return new Promise((resolve) => {
-    const req = indexedDB.deleteDatabase('habitat')
-    req.onsuccess = () => resolve()
-    req.onerror = () => resolve() // best-effort
-    req.onblocked = () => resolve()
-  })
-}
-
-function clearLocalStorage() {
-  for (const key of Object.keys(localStorage)) {
-    if (key.startsWith('checkin-')) localStorage.removeItem(key)
-  }
-}
-
-// ─── Shared wipe logic ───────────────────────────────────────────────────────
-
-/** True after a successful wipe-without-reload; both modals show a done state. */
-const wiped = ref(false)
-
-/**
- * Full wipe: closes the SQLite DB, removes all OPFS files, clears IDB and
- * localStorage. If reload=true the page reloads; otherwise sets wiped=true so
- * the modal can show a "done — close this tab" state.
- */
-async function fullWipe(reload: boolean): Promise<void> {
-  localStorage.removeItem('habitat-has-data')
-  await db.nukeOpfs() // worker closes DB + removes every OPFS entry
-  await clearIdb()
-  clearLocalStorage()
-  // journal entries live in localStorage under journal-YYYY-MM-DD keys
-  for (const key of Object.keys(localStorage)) {
-    if (key.startsWith('journal-')) localStorage.removeItem(key)
-  }
-  if (reload) {
-    window.location.reload()
-  } else {
-    wiped.value = true
-  }
-}
-
-// ─── Health setup ─────────────────────────────────────────────────────────────
-
-const showHealthSetup = ref(false)
-const creatingHealth = ref(false)
-
-const healthSetup = reactive({
-  enableSteps: true,
-  stepGoal: 10000,
-  enableMeals: true,
-  meals: [
-    { name: 'Breakfast', calories: 600 },
-    { name: 'Lunch', calories: 600 },
-    { name: 'Dinner', calories: 600 },
-  ] as { name: string; calories: number }[],
-  enableWater: true,
-  waterGoal: 8,
-  enableSleep: true,
-  sleepGoal: 8,
-})
-
-function addMeal() {
-  healthSetup.meals.push({ name: '', calories: 600 })
-}
-function removeMeal(i: number) {
-  healthSetup.meals.splice(i, 1)
-}
-
-async function onHealthToggle(value: boolean) {
-  setAppSetting('enableHealth', value)
-  if (value && db.isAvailable) {
-    const allHabits = await db.getHabits()
-    if (!allHabits.some((h) => h.tags.includes('habitat-health'))) {
-      showHealthSetup.value = true
-    }
-  }
-}
-
-async function confirmHealthSetup() {
-  if (!db.isAvailable) return
-  creatingHealth.value = true
-  try {
-    const base = { description: '', frequency: 'daily', annotations: {}, paused_until: null }
-    if (healthSetup.enableSteps) {
-      await db.createHabit({
-        ...base,
-        name: 'Steps',
-        type: 'NUMERIC',
-        target_value: healthSetup.stepGoal,
-        color: '#ef4444',
-        icon: 'i-heroicons-fire',
-        tags: ['habitat-health', 'habitat-steps'],
-      })
-    }
-    if (healthSetup.enableMeals) {
-      for (const meal of healthSetup.meals) {
-        if (!meal.name.trim()) continue
-        await db.createHabit({
-          ...base,
-          name: meal.name.trim(),
-          type: 'LIMIT',
-          target_value: meal.calories,
-          color: '#f59e0b',
-          icon: 'i-heroicons-sparkles',
-          tags: ['habitat-health', 'habitat-meals'],
-        })
-      }
-    }
-    if (healthSetup.enableWater) {
-      await db.createHabit({
-        ...base,
-        name: 'Water',
-        type: 'NUMERIC',
-        target_value: healthSetup.waterGoal,
-        color: '#0ea5e9',
-        icon: 'i-heroicons-beaker',
-        tags: ['habitat-health', 'habitat-water'],
-      })
-    }
-    if (healthSetup.enableSleep) {
-      await db.createHabit({
-        ...base,
-        name: 'Sleep',
-        type: 'NUMERIC',
-        target_value: healthSetup.sleepGoal,
-        color: '#6366f1',
-        icon: 'i-heroicons-moon',
-        tags: ['habitat-health', 'habitat-sleep'],
-      })
-    }
-    showHealthSetup.value = false
-  } finally {
-    creatingHealth.value = false
-  }
-}
-
-// ─── Variant 1: clear app data (selective, no OPFS wipe) ─────────────────────
-
-const showClearModal = ref(false)
-const clearing = ref(false)
-
-const clearSelection = reactive({
-  habits: true,
-  checkinEntries: true,
-  checkins: true,
-  scribbles: true,
-  voiceNotes: true,
-  todos: false,
-  boredData: false,
-  appliedDefaults: false,
-})
-
-const clearItems = [
-  {
-    key: 'habits',
-    label: 'Habits & completions',
-    description: 'All habits, logs, and completion history',
-  },
-  {
-    key: 'checkinEntries',
-    label: 'Check-in entries',
-    description: 'Daily check-in answers stored in the DB',
-  },
-  {
-    key: 'checkins',
-    label: 'Check-in templates & responses',
-    description: 'All templates, questions, and recorded answers',
-  },
-  { key: 'scribbles', label: 'Scribbles', description: 'All free-form notes' },
-  { key: 'voiceNotes', label: 'Voice recordings', description: 'IndexedDB audio data' },
-  { key: 'todos', label: 'TODOs', description: 'All tasks and their history' },
-  {
-    key: 'boredData',
-    label: 'Bored activities',
-    description: 'All custom activities (system categories preserved)',
-  },
-  {
-    key: 'appliedDefaults',
-    label: 'Applied defaults',
-    description: 'Re-enables seeding of default check-in templates',
-  },
-] as const
-
-const nothingSelected = computed(
-  () =>
-    !(
-      clearSelection.habits ||
-      clearSelection.checkinEntries ||
-      clearSelection.checkins ||
-      clearSelection.scribbles ||
-      clearSelection.voiceNotes ||
-      clearSelection.todos ||
-      clearSelection.boredData ||
-      clearSelection.appliedDefaults
-    ),
-)
-
-async function clearAppData() {
-  if (!db.isAvailable || nothingSelected.value) return
-  clearing.value = true
-  try {
-    const ops: Promise<unknown>[] = []
-    if (clearSelection.habits) ops.push(db.deleteAllHabits())
-    if (clearSelection.checkinEntries) ops.push(db.deleteAllCheckinEntries())
-    if (clearSelection.checkins) ops.push(db.deleteAllCheckinData())
-    if (clearSelection.scribbles) ops.push(db.deleteAllScribbles())
-    if (clearSelection.todos) ops.push(db.deleteAllTodos())
-    if (clearSelection.boredData) ops.push(db.deleteAllBoredData())
-    if (clearSelection.appliedDefaults) ops.push(db.clearAppliedDefaults())
-    await Promise.all(ops)
-    if (clearSelection.checkinEntries) clearLocalStorage()
-    if (clearSelection.voiceNotes) await clearIdb()
-    if (clearSelection.habits) localStorage.removeItem('habitat-has-data')
-    showClearModal.value = false
-  } finally {
-    clearing.value = false
-  }
-}
-
-// ─── Variant 2: wipe all OPFS files (nuclear) ────────────────────────────────
-
-const showNukeModal = ref(false)
-const nuking = ref(false)
-
-async function nukeOpfs(reload: boolean) {
-  if (!db.isAvailable) return
-  nuking.value = true
-  try {
-    await fullWipe(reload)
-  } catch {
-    showNukeModal.value = false
-  } finally {
-    nuking.value = false
-  }
-}
-
-// ─── Advanced: integrity check ────────────────────────────────────────────────
-
-const integrityResults = ref<string[] | null>(null)
-const integrityLoading = ref(false)
-
-async function runIntegrityCheck() {
-  if (!db.isAvailable) return
-  integrityLoading.value = true
-  integrityResults.value = null
-  try {
-    integrityResults.value = await db.integrityCheck()
-  } finally {
-    integrityLoading.value = false
-  }
-}
-
-const integrityOk = computed(
-  () =>
-    integrityResults.value !== null &&
-    integrityResults.value.length === 1 &&
-    integrityResults.value[0] === 'ok',
-)
-
-// ─── Advanced: DB info ────────────────────────────────────────────────────────
-
-import { Capacitor } from '@capacitor/core'
-import { zipSync } from 'fflate'
-import type { AppTheme } from '~/composables/useAppSettings'
-import type { DbInfo, ExportSelection, HabitatExport } from '~/types/database'
-
-// ── Appearance ────────────────────────────────────────────────────────────────
-
-const colorMode = useColorMode()
-
-const THEMES: { id: AppTheme; name: string; swatch: string }[] = [
-  { id: 'habitat', name: 'Habitat', swatch: '#22d3ee' },
-  { id: 'forest', name: 'Forest', swatch: '#208a65' },
-  { id: 'ocean', name: 'Ocean', swatch: '#6366f1' },
-]
-
-function setTheme(theme: AppTheme) {
-  if (!import.meta.client) return
-  document.documentElement.classList.add('theme-transitioning')
-  setAppSetting('theme', theme)
-  setTimeout(() => document.documentElement.classList.remove('theme-transitioning'), 250)
-}
-
-const dbInfoOpen = ref(false)
-const dbInfo = ref<DbInfo | null>(null)
-const dbInfoLoading = ref(false)
-const expandedTable = ref<string | null>(null)
-
-async function loadDbInfo() {
-  if (!db.isAvailable) return
-  dbInfoLoading.value = true
-  try {
-    dbInfo.value = await db.getDbInfo()
-  } finally {
-    dbInfoLoading.value = false
-  }
-}
-
-function toggleDbInfo() {
-  dbInfoOpen.value = !dbInfoOpen.value
-  if (dbInfoOpen.value && !dbInfo.value) loadDbInfo()
-}
-
-// ─── Advanced: OPFS files ─────────────────────────────────────────────────────
-
-type OpfsFile = { path: string; size: number }
-
-const opfsOpen = ref(false)
-const opfsFiles = ref<OpfsFile[]>([])
-const opfsLoading = ref(false)
-
-async function walkOpfs(dir: FileSystemDirectoryHandle, prefix = ''): Promise<OpfsFile[]> {
-  const out: OpfsFile[] = []
-  // @ts-expect-error — entries() not in TypeScript's FileSystemDirectoryHandle types
-  for await (const [name, handle] of dir.entries()) {
-    const path = prefix ? `${prefix}/${name}` : name
-    if ((handle as FileSystemHandle).kind === 'directory') {
-      out.push(...(await walkOpfs(handle as FileSystemDirectoryHandle, path)))
-    } else {
-      const file = await (handle as FileSystemFileHandle).getFile()
-      out.push({ path, size: file.size })
-    }
-  }
-  return out
-}
-
-async function loadOpfsFiles() {
-  opfsLoading.value = true
-  try {
-    const root = await navigator.storage.getDirectory()
-    opfsFiles.value = await walkOpfs(root)
-  } finally {
-    opfsLoading.value = false
-  }
-}
-
-function toggleOpfs() {
-  opfsOpen.value = !opfsOpen.value
-  if (opfsOpen.value) loadOpfsFiles()
-}
-
-function formatBytes(n: number): string {
-  if (n < 1024) return `${n} B`
-  if (n < 1024 * 1024) return `${(n / 1024).toFixed(1)} KB`
-  return `${(n / (1024 * 1024)).toFixed(2)} MB`
-}
-
-// ─── Advanced: force reload ───────────────────────────────────────────────────
-
-const forceReloading = ref(false)
-
-async function forceReload() {
-  forceReloading.value = true
-  try {
-    // Unregister every SW for this origin so the browser installs a fresh one
-    if ('serviceWorker' in navigator) {
-      const registrations = await navigator.serviceWorker.getRegistrations()
-      await Promise.all(registrations.map((r) => r.unregister()))
-    }
-    // Clear all Cache API caches (SW asset caches) — OPFS is unaffected
-    if ('caches' in window) {
-      const keys = await caches.keys()
-      await Promise.all(keys.map((k) => caches.delete(k)))
-    }
-    window.location.reload()
-  } catch {
-    forceReloading.value = false
-  }
-}
-
-// ─── Install ──────────────────────────────────────────────────────────────────
-
-const { isInstalled, canInstall, isIosSafari, isChromiumNoPrompt, installing, install } =
-  useInstall()
-const isNativeApp = Capacitor.isNativePlatform()
-const isAndroid = Capacitor.getPlatform() === 'android'
-const isIos = isNativeApp && !isAndroid
-const showInstallModal = ref(false)
-
-function handleInstall() {
-  if (canInstall.value) {
-    install()
-  } else {
-    showInstallModal.value = true
-  }
-}
-
-// ─── Notifications & Permissions ──────────────────────────────────────────────
-
-const {
-  permission: notifPermission,
-  exactAlarm,
-  batteryOptim,
-  persistentStorage,
-  notifLog,
-  requestPermission,
-  openExactAlarmSetting,
-  requestBatteryExemption,
-  checkExactAlarm,
-  checkBatteryOptim,
-  checkPersistentStorage,
-  requestPersistentStorage,
-  sendTestNotification,
-  testScheduleOn,
-} = useNotifications()
-
-async function enableNotifications() {
-  await requestPermission()
-}
-
-onMounted(async () => {
-  if (isNativeApp) {
-    await checkExactAlarm()
-    await checkBatteryOptim()
-  } else {
-    await checkPersistentStorage()
-  }
-  await checkMediaPermissions()
-})
-
-// ─── Licenses ─────────────────────────────────────────────────────────────────
-
-interface LicenseEntry {
-  name: string
-  version: string
-  license: string | null
-  homepage: string | null
-}
-const showLicensesModal = ref(false)
-const licenses = ref<LicenseEntry[]>([])
-const licensesLoading = ref(false)
-const licensesError = ref(false)
-
-async function openLicenses() {
-  showLicensesModal.value = true
-  if (licenses.value.length > 0) return
-  licensesLoading.value = true
-  licensesError.value = false
-  try {
-    const base = useRuntimeConfig().app.baseURL
-    const data = await $fetch<LicenseEntry[]>(`${base}licenses.json`)
-    licenses.value = data
-  } catch {
-    licensesError.value = true
-  } finally {
-    licensesLoading.value = false
-  }
-}
-
-// ─── Advanced: storage estimate ───────────────────────────────────────────────
-
-const storageEstimate = ref<StorageEstimate | null>(null)
-const storagePersisted = ref<boolean | null>(null)
-
-async function loadStorageEstimate() {
-  if (!('storage' in navigator)) return
-  if ('estimate' in navigator.storage) {
-    storageEstimate.value = await navigator.storage.estimate()
-  }
-  if ('persisted' in navigator.storage) {
-    storagePersisted.value = await navigator.storage.persisted()
-  }
-}
-
-async function requestPersistStorage() {
-  if (!('storage' in navigator && 'persist' in navigator.storage)) return
-  storagePersisted.value = await navigator.storage.persist()
-}
-
-// ─── Microphone / Camera permissions ─────────────────────────────────────────
-
-const micPermission = ref<PermissionState | null>(null)
-const cameraPermission = ref<PermissionState | null>(null)
-const photoLibraryPermission = ref<PermissionState | null>(null)
-
-async function checkMediaPermissions() {
-  if (!navigator.permissions) return
-  for (const [name, target] of [
-    ['microphone', micPermission],
-    ['camera', cameraPermission],
-  ] as const) {
-    try {
-      const status = await navigator.permissions.query({ name: name as PermissionName })
-      target.value = status.state
-      status.onchange = () => {
-        target.value = status.state
-      }
-    } catch {
-      // Not supported in this browser (e.g. Firefox for camera)
-    }
-  }
-  if (isIos) {
-    try {
-      // 'photos' is not a standard name — will throw on most platforms; result stays null (Unknown)
-      const status = await navigator.permissions.query({ name: 'photos' as PermissionName })
-      photoLibraryPermission.value = status.state
-    } catch {
-      // Expected: shown as Unknown
-    }
-  }
-}
-
-// ─── Strict CSP ───────────────────────────────────────────────────────────────
-
-// Was the strict CSP already active when this page loaded?
-// Set by strict-csp.client.ts plugin at startup; false before plugin runs.
-const strictCspSessionActive = useState<boolean>('strict-csp-session-active', () => false)
-
-function enableStrictCspAndReload(value: boolean) {
-  if (!value) return // toggle can only go from off → on this way; disable uses its own button
-  setAppSetting('strictCsp', true)
-  window.location.reload()
-}
-
-function disableStrictCspAndReload() {
-  setAppSetting('strictCsp', false)
-  window.location.reload()
-}
-
-// ─── Collapsible section state ────────────────────────────────────────────────
-
-const aboutOpen = ref(false)
-const diagOpen = ref(false)
-const dragonsOpen = ref(false)
-const showNotifLog = ref(false)
-
-// ─── Pending notifications (diagnostics) ──────────────────────────────────────
-
-const { scheduled: scheduledNotifs, listPending: loadPendingNotifs } = useNotifications()
-const pendingNotifsLoading = ref(false)
-
-async function refreshPendingNotifs() {
-  pendingNotifsLoading.value = true
-  try {
-    await loadPendingNotifs()
-  } finally {
-    pendingNotifsLoading.value = false
-  }
-}
-
-watch(diagOpen, (open) => {
-  if (open) {
-    loadStorageEstimate()
-    void refreshPendingNotifs()
-  }
-})
 </script>
 
 <template>
-  <div class="space-y-6">
-    <header>
-      <p class="text-sm text-(--ui-text-dimmed)">Preferences</p>
-      <h2 class="text-2xl font-bold">Settings</h2>
-    </header>
-
-    <!-- ── Display ───────────────────────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed) px-1">Display</p>
-      <UCard :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <!-- Appearance subsection -->
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Appearance</p>
-        </div>
-
-        <!-- Theme swatches -->
-        <div class="flex items-center justify-between px-4 py-3">
-          <p class="text-sm font-medium">Theme</p>
-          <div class="flex gap-2.5 items-center">
-            <button
-              v-for="t in THEMES"
-              :key="t.id"
-              class="w-8 h-8 rounded-full transition-transform focus:outline-none focus-visible:ring-2 focus-visible:ring-primary-500"
-              :class="appSettings.theme === t.id
-                ? 'ring-2 ring-offset-2 ring-offset-(--ui-bg-muted) ring-primary-500 scale-110'
-                : 'opacity-70 hover:opacity-100 hover:scale-105'"
-              :style="{ background: t.swatch }"
-              :title="t.name"
-              :aria-label="`${t.name} theme`"
-              :aria-pressed="appSettings.theme === t.id"
-              @click="setTheme(t.id)"
-            />
-          </div>
-        </div>
-
-        <!-- Color mode toggle -->
-        <div class="flex items-center justify-between px-4 py-3">
-          <p class="text-sm font-medium">Color mode</p>
-          <div class="flex bg-(--ui-bg-elevated) rounded-lg p-0.5 gap-0.5">
-            <button
-              class="px-3 py-1 rounded-md text-xs font-medium transition-colors"
-              :class="colorMode.value !== 'dark' ? 'bg-(--ui-bg) text-(--ui-text) shadow-sm' : 'text-(--ui-text-dimmed) hover:text-(--ui-text-toned)'"
-              @click="colorMode.preference = 'light'"
-            >Light</button>
-            <button
-              class="px-3 py-1 rounded-md text-xs font-medium transition-colors"
-              :class="colorMode.value === 'dark' ? 'bg-(--ui-bg) text-(--ui-text) shadow-sm' : 'text-(--ui-text-dimmed) hover:text-(--ui-text-toned)'"
-              @click="colorMode.preference = 'dark'"
-            >Dark</button>
-          </div>
-        </div>
-
-        <!-- Reduce motion -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Reduce motion</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Disables animations and transitions.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.reduceMotion"
-            @update:model-value="setAppSetting('reduceMotion', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">24-hour time</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Show times as 17:34 instead of 5:34 PM.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.use24HourTime"
-            @update:model-value="setAppSetting('use24HourTime', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Sticky bottom bar</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Keep the nav bar fixed at the bottom when scrolling.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.stickyNav"
-            @update:model-value="setAppSetting('stickyNav', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Extra bottom padding</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Add space below the nav for Android gesture buttons.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.navExtraPadding"
-            @update:model-value="setAppSetting('navExtraPadding', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Extra top padding</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Add space above the header for the status bar.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.headerExtraPadding"
-            @update:model-value="setAppSetting('headerExtraPadding', $event)"
-          />
-        </div>
-
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Navigation</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Today tab</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Show daily progress and log habits for today.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableToday"
-            @update:model-value="setAppSetting('enableToday', $event)"
-          />
-        </div>
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Matrix view</p>
-        </div>
-        <div class="flex items-center justify-between px-4 pb-3">
-          <p class="text-sm text-(--ui-text-muted)">Days to show (mobile)</p>
-          <div class="flex items-center gap-2">
-            <button
-              class="w-7 h-7 rounded-lg bg-(--ui-bg-elevated) border border-(--ui-border-accented) text-(--ui-text-toned) flex items-center justify-center text-sm"
-              @click="setAppSetting('weekDays', Math.max(3, (appSettings.weekDays || 3) - 1))"
-            >−</button>
-            <span class="w-4 text-center text-sm font-medium tabular-nums">{{ appSettings.weekDays || 3 }}</span>
-            <button
-              class="w-7 h-7 rounded-lg bg-(--ui-bg-elevated) border border-(--ui-border-accented) text-(--ui-text-toned) flex items-center justify-center text-sm"
-              @click="setAppSetting('weekDays', Math.min(7, (appSettings.weekDays || 3) + 1))"
-            >+</button>
-          </div>
-        </div>
-        <div class="flex items-center justify-between px-4 pb-3">
-          <p class="text-sm text-(--ui-text-muted)">Newest day first</p>
-          <USwitch
-            :model-value="appSettings.matrixReverseDays"
-            @update:model-value="setAppSetting('matrixReverseDays', $event)"
-          />
-        </div>
-
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Habits page</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Show tags</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Display habit tags in the habits list.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.showTagsOnHabits"
-            @update:model-value="setAppSetting('showTagsOnHabits', $event)"
-          />
-        </div>
-        <div class="flex items-center justify-between px-4 py-3">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Show annotations</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Display key:value metadata in the habits list.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.showAnnotationsOnHabits"
-            @update:model-value="setAppSetting('showAnnotationsOnHabits', $event)"
-          />
-        </div>
-
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Today page</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Show tags</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Display habit tags in today's list.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.showTagsOnToday"
-            @update:model-value="setAppSetting('showTagsOnToday', $event)"
-          />
-        </div>
-        <div class="flex items-center justify-between px-4 py-3">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Show annotations</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Display key:value metadata in today's list.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.showAnnotationsOnToday"
-            @update:model-value="setAppSetting('showAnnotationsOnToday', $event)"
-          />
-        </div>
-
-        <div class="px-4 pt-3 pb-1">
-          <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600">Log entries</p>
-        </div>
-        <div class="px-4 py-3 pb-4 space-y-2">
-          <p class="text-sm font-medium">Input mode</p>
-          <p class="text-xs text-(--ui-text-dimmed) mb-2">How numeric values are applied when logging a habit.</p>
-          <div class="flex gap-2">
-            <button
-              class="flex-1 py-2 px-3 rounded-xl text-xs font-medium border transition-colors"
-              :class="appSettings.logInputMode === 'increment'
-                ? 'bg-primary-500/15 border-primary-500/40 text-primary-300'
-                : 'bg-(--ui-bg-elevated) border-(--ui-border-accented) text-(--ui-text-muted)'"
-              @click="setAppSetting('logInputMode', 'increment')"
-            >
-              Add to total
-            </button>
-            <button
-              class="flex-1 py-2 px-3 rounded-xl text-xs font-medium border transition-colors"
-              :class="appSettings.logInputMode === 'absolute'
-                ? 'bg-primary-500/15 border-primary-500/40 text-primary-300'
-                : 'bg-(--ui-bg-elevated) border-(--ui-border-accented) text-(--ui-text-muted)'"
-              @click="setAppSetting('logInputMode', 'absolute')"
-            >
-              Set total
-            </button>
-          </div>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Features ──────────────────────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed) px-1">Features</p>
-      <UCard :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Enable Journalling</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Show Check-in, Scribbles, and Voice in the navigation.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableJournalling"
-            @update:model-value="setAppSetting('enableJournalling', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Enable Health</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Track daily steps and meal calories.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableHealth"
-            @update:model-value="onHealthToggle"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Enable TODOs</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Standalone task tracker with due dates, priorities, and recurring tasks.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableTodos"
-            @update:model-value="setAppSetting('enableTodos', $event)"
-          />
-        </div>
-
-        <div v-if="appSettings.enableTodos" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Enable "I'm Bored" Mode</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Magic 8-ball oracle that suggests activities from curated categories.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableBored"
-            @update:model-value="setAppSetting('enableBored', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Context Filter</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Tag icon in the header lets you filter habits, todos, and bored suggestions by shared tags.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableContextFilter"
-            @update:model-value="setAppSetting('enableContextFilter', $event)"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Timer & Focus</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Start button on TODOs and Bored activities with stopwatch, countdown, and Pomodoro modes.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.enableTimer"
-            @update:model-value="setAppSetting('enableTimer', $event)"
-          />
-        </div>
-
-        <div v-if="appSettings.enableJournalling" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Save transcriptions</p>
-            <p class="text-xs text-(--ui-text-dimmed)">After recording, offer to save the speech-to-text transcript as a Scribble tagged <code class="text-(--ui-text-muted)">habitat-transcribed</code>.</p>
-          </div>
-          <USwitch
-            :model-value="appSettings.saveTranscribedNotes"
-            @update:model-value="setAppSetting('saveTranscribedNotes', $event)"
-          />
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Pomodoro ──────────────────────────────────────────────────────────── -->
-    <section v-if="appSettings.enableTimer" class="space-y-2">
-      <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed) px-1">Pomodoro</p>
-      <UCard :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm font-medium">Work block</p>
-          <div class="flex items-center gap-1.5">
-            <input
-              type="number"
-              min="1"
-              max="90"
-              :value="appSettings.pomodoroWorkMinutes"
-              class="w-14 px-2 py-1 text-sm bg-(--ui-bg-elevated) border border-(--ui-border) rounded-lg text-center"
-              @change="setAppSetting('pomodoroWorkMinutes', +($event.target as HTMLInputElement).value)"
-            />
-            <span class="text-xs text-(--ui-text-dimmed)">min</span>
-          </div>
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm font-medium">Short break</p>
-          <div class="flex items-center gap-1.5">
-            <input
-              type="number"
-              min="1"
-              max="60"
-              :value="appSettings.pomodoroShortBreakMinutes"
-              class="w-14 px-2 py-1 text-sm bg-(--ui-bg-elevated) border border-(--ui-border) rounded-lg text-center"
-              @change="setAppSetting('pomodoroShortBreakMinutes', +($event.target as HTMLInputElement).value)"
-            />
-            <span class="text-xs text-(--ui-text-dimmed)">min</span>
-          </div>
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm font-medium">Long break</p>
-          <div class="flex items-center gap-1.5">
-            <input
-              type="number"
-              min="1"
-              max="60"
-              :value="appSettings.pomodoroLongBreakMinutes"
-              class="w-14 px-2 py-1 text-sm bg-(--ui-bg-elevated) border border-(--ui-border) rounded-lg text-center"
-              @change="setAppSetting('pomodoroLongBreakMinutes', +($event.target as HTMLInputElement).value)"
-            />
-            <span class="text-xs text-(--ui-text-dimmed)">min</span>
-          </div>
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm font-medium">Cycles before long break</p>
-          <div class="flex items-center gap-1.5">
-            <input
-              type="number"
-              min="1"
-              max="10"
-              :value="appSettings.pomodoroCyclesBeforeLong"
-              class="w-14 px-2 py-1 text-sm bg-(--ui-bg-elevated) border border-(--ui-border) rounded-lg text-center"
-              @change="setAppSetting('pomodoroCyclesBeforeLong', +($event.target as HTMLInputElement).value)"
-            />
-          </div>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Permissions ─────────────────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed) px-1">Permissions</p>
-      <UCard :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <!-- Install -->
-        <div v-if="isInstalled" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-arrow-down-tray" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Habitat is installed</p>
-            </div>
-            <p class="text-xs text-green-400">{{ isNativeApp ? 'Running as a native app' : 'Running as a standalone app' }}</p>
-          </div>
-          <div class="w-2 h-2 rounded-full bg-green-400 mx-2 shrink-0" />
-        </div>
-        <div v-else class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-arrow-down-tray" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Install Habitat</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Add to your home screen for offline access and notifications</p>
-          </div>
-          <UButton
-            size="sm"
-            variant="soft"
-            color="primary"
-            icon="i-heroicons-arrow-down-tray"
-            :loading="installing"
-            class="shrink-0"
-            @click="handleInstall"
-          >
-            Install
-          </UButton>
-        </div>
-
-        <!-- Notifications -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-bell" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Notifications</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Deliver reminders for habits and check-ins.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="notifPermission === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="notifPermission === 'denied'"
-              label="Blocked" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="isIosSafari"
-              label="Unavailable" variant="subtle" color="warning" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Required" variant="subtle" color="warning" size="xs" class="rounded-full"
-            />
-            <UButton
-              v-if="notifPermission === 'default' && !isIosSafari"
-              size="xs" variant="soft" color="primary"
-              @click="enableNotifications"
-            >
-              Enable
-            </UButton>
-          </div>
-        </div>
-
-        <!-- Exact Alarms (Android native only) -->
-        <div v-if="isNativeApp" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-clock" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Exact alarms</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Fire reminders at the exact time you scheduled them.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="exactAlarm === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="exactAlarm === 'denied'"
-              label="Denied" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Unknown" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UButton
-              v-if="exactAlarm === 'denied'"
-              size="xs" variant="soft" color="warning"
-              @click="openExactAlarmSetting"
-            >
-              Fix
-            </UButton>
-          </div>
-        </div>
-
-        <!-- Battery Optimization (Android native only) -->
-        <div v-if="isNativeApp" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-battery-100" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Battery optimization</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Exempt the app so reminders fire even when closed.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="batteryOptim === 'exempt'"
-              label="Exempt" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="batteryOptim === 'optimized'"
-              label="Optimized" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Unknown" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UButton
-              v-if="batteryOptim === 'optimized'"
-              size="xs" variant="soft" color="warning"
-              @click="requestBatteryExemption"
-            >
-              Fix
-            </UButton>
-          </div>
-        </div>
-
-        <!-- Persistent Storage (web PWA only) -->
-        <div v-if="!isNativeApp" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon
-                :name="persistentStorage === 'granted' ? 'i-heroicons-lock-closed' : 'i-heroicons-lock-open'"
-                class="w-4 h-4 shrink-0"
-                :class="persistentStorage === 'granted' ? 'text-green-400' : 'text-(--ui-text-muted)'"
-              />
-              <p class="text-sm font-medium">Persistent storage</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Prevent the browser from evicting your offline data.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="persistentStorage === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="persistentStorage === 'denied'"
-              label="Not persisted" variant="subtle" color="warning" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Unknown" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UButton
-              v-if="persistentStorage === 'denied'"
-              size="xs" variant="soft" color="primary"
-              @click="() => { requestPersistentStorage() }"
-            >
-              Request
-            </UButton>
-          </div>
-        </div>
-
-        <!-- Microphone -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-microphone" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Microphone</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Required for voice recordings in Jots.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="micPermission === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="micPermission === 'denied'"
-              label="Blocked" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="micPermission === 'prompt'"
-              label="Not yet asked" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Unknown" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-          </div>
-        </div>
-
-        <!-- Camera -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-camera" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Camera</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Required to capture photos in Jots.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="cameraPermission === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="cameraPermission === 'denied'"
-              label="Blocked" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="cameraPermission === 'prompt'"
-              label="Not yet asked" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Unknown" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-          </div>
-        </div>
-
-        <!-- Photo Library (iOS native only) -->
-        <div v-if="isIos" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 min-w-0">
-            <div class="flex items-center gap-2">
-              <UIcon name="i-heroicons-photo" class="w-4 h-4 text-(--ui-text-muted) shrink-0" />
-              <p class="text-sm font-medium">Photo library</p>
-            </div>
-            <p class="text-xs text-(--ui-text-dimmed)">Required to pick existing photos for Jots.</p>
-          </div>
-          <div class="shrink-0 flex items-center gap-2">
-            <UBadge
-              v-if="photoLibraryPermission === 'granted'"
-              label="Granted" variant="subtle" color="success" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="photoLibraryPermission === 'denied'"
-              label="Blocked" variant="subtle" color="error" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else-if="photoLibraryPermission === 'prompt'"
-              label="Not yet asked" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-            <UBadge
-              v-else
-              label="Managed in Settings" variant="subtle" color="neutral" size="xs" class="rounded-full"
-            />
-          </div>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Data ──────────────────────────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed) px-1">Data</p>
-      <UCard :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Export data</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Download selected data as a versioned JSON file.</p>
-          </div>
-          <UButton
-            icon="i-heroicons-arrow-down-tray"
-            variant="ghost"
-            color="neutral"
-            size="sm"
-            :disabled="!db.isAvailable"
-            @click="openExportModal"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Import data</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Merge from a Habitat JSON export. Existing records are kept.</p>
-          </div>
-          <input ref="importInput" type="file" accept=".json" class="hidden" @change="onImportFileSelected" />
-          <UButton
-            icon="i-heroicons-arrow-up-tray"
-            variant="ghost"
-            color="neutral"
-            size="sm"
-            :disabled="!db.isAvailable"
-            @click="openImport"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Export SQLite</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Download the raw <span class="font-mono">.sqlite3</span> database file.</p>
-          </div>
-          <UButton
-            icon="i-heroicons-circle-stack"
-            variant="ghost"
-            color="neutral"
-            size="sm"
-            :loading="exportingDb"
-            :disabled="!db.isAvailable"
-            @click="exportSqlite"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Export Jots</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Download text notes, voice recordings, and images as a <span class="font-mono">.zip</span>.</p>
-          </div>
-          <UButton
-            icon="i-heroicons-document-text"
-            variant="ghost"
-            color="neutral"
-            size="sm"
-            @click="showJotsExportModal = true"
-          />
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── About (collapsible) ───────────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <button class="w-full flex items-center justify-between px-1 py-0.5" @click="aboutOpen = !aboutOpen">
-        <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed)">About</p>
-        <UIcon :name="aboutOpen ? 'i-heroicons-chevron-up' : 'i-heroicons-chevron-down'" class="w-3.5 h-3.5 text-slate-600" />
-      </button>
-      <UCard v-if="aboutOpen" :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">App</p>
-          <p class="text-sm font-medium">Habitat</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Version</p>
-          <p class="text-sm font-mono font-medium">{{ runtimeConfig.public.appVersion || '—' }}</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Commit</p>
-          <p class="text-sm font-mono text-(--ui-text-toned)">{{ runtimeConfig.public.commitSha || '—' }}</p>
-        </div>
-        <div v-if="runtimeConfig.public.gitTag.length > 0" class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Tag</p>
-          <UBadge :label="runtimeConfig.public.gitTag" variant="subtle" color="primary" size="sm" class="rounded-full font-mono" />
-        </div>
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Build</p>
-          <UBadge :label="runtimeConfig.public.buildTarget" variant="subtle" color="neutral" size="sm" class="rounded-full font-mono" />
-        </div>
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Built</p>
-          <p class="text-sm font-mono text-(--ui-text-toned)">{{ runtimeConfig.public.buildTime ? new Date(runtimeConfig.public.buildTime).toLocaleString() : '—' }}</p>
-        </div>
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <p class="text-sm text-(--ui-text-muted)">Storage</p>
-          <p class="text-sm font-medium">On-device (OPFS)</p>
-        </div>
-        <button class="w-full flex items-center justify-between px-4 py-3.5 text-left" @click="openLicenses">
-          <p class="text-sm text-(--ui-text-muted)">Open source licenses</p>
-          <UIcon name="i-heroicons-chevron-right" class="w-4 h-4 text-(--ui-text-dimmed) shrink-0" />
-        </button>
-
-        <!-- DB schema -->
-        <div>
-          <button class="w-full flex items-center justify-between px-4 py-3.5 text-left" @click="toggleDbInfo">
-            <div class="space-y-0.5">
-              <p class="text-sm text-(--ui-text-muted)">Database schema</p>
-            </div>
-            <div class="flex items-center gap-2 shrink-0">
-              <UBadge v-if="dbInfo" :label="`v${dbInfo.userVersion}`" variant="subtle" color="neutral" size="sm" class="font-mono rounded-full" />
-              <UIcon :name="dbInfoOpen ? 'i-heroicons-chevron-down' : 'i-heroicons-chevron-right'" class="w-4 h-4 text-(--ui-text-dimmed)" />
-            </div>
-          </button>
-          <div v-if="dbInfoOpen" class="border-t border-(--ui-border) px-4 py-3 space-y-2">
-            <div v-if="dbInfoLoading" class="flex items-center gap-2 text-xs text-(--ui-text-dimmed)">
-              <UIcon name="i-heroicons-arrow-path" class="w-3.5 h-3.5 animate-spin" /> Loading…
-            </div>
-            <template v-else-if="dbInfo">
-              <div class="flex items-center gap-2 mb-3">
-                <span class="text-xs text-(--ui-text-dimmed)">Schema version:</span>
-                <span class="font-mono text-xs text-(--ui-text)">{{ dbInfo.userVersion }}</span>
-              </div>
-              <div v-for="table in dbInfo.tables" :key="table.name" class="rounded-lg border border-(--ui-border) overflow-hidden">
-                <button
-                  class="w-full flex items-center justify-between px-3 py-2 text-left hover:bg-(--ui-bg-elevated)/40 transition-colors"
-                  @click="expandedTable = expandedTable === table.name ? null : table.name"
-                >
-                  <span class="text-xs font-mono text-(--ui-text-toned)">{{ table.name }}</span>
-                  <UIcon :name="expandedTable === table.name ? 'i-heroicons-chevron-down' : 'i-heroicons-chevron-right'" class="w-3.5 h-3.5 text-slate-600" />
-                </button>
-                <pre v-if="expandedTable === table.name" class="text-[10px] leading-relaxed font-mono text-(--ui-text-muted) bg-(--ui-bg) px-3 py-2 overflow-x-auto border-t border-(--ui-border)">{{ table.sql }}</pre>
-              </div>
-              <template v-if="dbInfo.indices.length > 0">
-                <p class="text-[10px] font-semibold uppercase tracking-wider text-slate-600 pt-1">Indices</p>
-                <div v-for="idx in dbInfo.indices" :key="idx.name" class="rounded-lg border border-(--ui-border) overflow-hidden">
-                  <button
-                    class="w-full flex items-center justify-between px-3 py-2 text-left hover:bg-(--ui-bg-elevated)/40 transition-colors"
-                    @click="expandedTable = expandedTable === idx.name ? null : idx.name"
-                  >
-                    <div class="min-w-0 text-left">
-                      <span class="text-xs font-mono text-(--ui-text-muted)">{{ idx.name }}</span>
-                      <span class="text-[10px] text-slate-600 ml-2">on {{ idx.tbl_name }}</span>
-                    </div>
-                    <UIcon :name="expandedTable === idx.name ? 'i-heroicons-chevron-down' : 'i-heroicons-chevron-right'" class="w-3.5 h-3.5 text-slate-600 shrink-0" />
-                  </button>
-                  <pre v-if="expandedTable === idx.name" class="text-[10px] leading-relaxed font-mono text-(--ui-text-muted) bg-(--ui-bg) px-3 py-2 overflow-x-auto border-t border-(--ui-border)">{{ idx.sql }}</pre>
-                </div>
-              </template>
-            </template>
-          </div>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Diagnostics (collapsible) ─────────────────────────────────────────── -->
-    <section class="space-y-2">
-      <button class="w-full flex items-center justify-between px-1 py-0.5" @click="diagOpen = !diagOpen">
-        <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed)">Diagnostics</p>
-        <UIcon :name="diagOpen ? 'i-heroicons-chevron-up' : 'i-heroicons-chevron-down'" class="w-3.5 h-3.5 text-slate-600" />
-      </button>
-      <UCard v-if="diagOpen" :ui="{ root: 'rounded-2xl', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <!-- Integrity check -->
-        <div class="flex items-start justify-between px-4 py-3.5 gap-3">
-          <div class="space-y-0.5 min-w-0">
-            <p class="text-sm font-medium">Integrity check</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Runs SQLite <span class="font-mono">PRAGMA integrity_check</span></p>
-            <div v-if="integrityResults !== null" class="mt-1.5">
-              <p v-if="integrityOk" class="text-xs text-green-400 flex items-center gap-1">
-                <UIcon name="i-heroicons-check-circle" class="w-3.5 h-3.5 shrink-0" /> ok
-              </p>
-              <ul v-else class="space-y-0.5">
-                <li v-for="(msg, i) in integrityResults" :key="i" class="text-xs text-red-400 font-mono break-all">{{ msg }}</li>
-              </ul>
-            </div>
-          </div>
-          <span
-            class="shrink-0"
-            :class="!db.isAvailable ? 'cursor-not-allowed' : ''"
-            :title="!db.isAvailable ? 'Database not ready' : undefined"
-          >
-            <UButton
-              size="sm" variant="ghost" color="neutral"
-              :loading="integrityLoading" :disabled="!db.isAvailable"
-              icon="i-heroicons-shield-check"
-              @click="runIntegrityCheck"
-            />
-          </span>
-        </div>
-
-        <!-- Test notification -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Test notification (instant)</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Fire via schedule.at — confirms the plugin works.</p>
-          </div>
-          <span
-            class="shrink-0"
-            :class="notifPermission !== 'granted' ? 'cursor-not-allowed' : ''"
-            :title="notifPermission !== 'granted' ? 'Notifications not granted' : undefined"
-          >
-            <UButton
-              size="sm" variant="ghost" color="neutral"
-              icon="i-heroicons-bell"
-              :disabled="notifPermission !== 'granted'"
-              @click="sendTestNotification"
-            />
-          </span>
-        </div>
-
-        <!-- Test schedule.on -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Test schedule.on (≈2 min)</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Schedules a repeating alarm for 2 min from now. Same API as real reminders.</p>
-          </div>
-          <span
-            class="shrink-0"
-            :class="notifPermission !== 'granted' || !isNativeApp ? 'cursor-not-allowed' : ''"
-            :title="!isNativeApp ? 'Native only' : notifPermission !== 'granted' ? 'Notifications not granted' : undefined"
-          >
-            <UButton
-              size="sm" variant="ghost" color="neutral"
-              icon="i-heroicons-clock"
-              :disabled="notifPermission !== 'granted' || !isNativeApp"
-              @click="testScheduleOn"
-            />
-          </span>
-        </div>
-
-        <!-- Notification log (gated) -->
-        <div>
-          <div class="flex items-center justify-between px-4 py-3.5">
-            <div class="space-y-0.5">
-              <p class="text-sm font-medium">Notification log</p>
-              <p class="text-xs text-(--ui-text-dimmed)">Recent notification events (scheduled, received, tapped).</p>
-            </div>
-            <div class="flex items-center gap-2 shrink-0">
-              <UBadge v-if="showNotifLog && notifLog.length > 0" :label="String(notifLog.length)" variant="subtle" color="neutral" size="xs" class="rounded-full font-mono" />
-              <USwitch :model-value="showNotifLog" @update:model-value="showNotifLog = $event" />
-            </div>
-          </div>
-          <div v-if="showNotifLog" class="border-t border-(--ui-border) px-4 pb-4 pt-3">
-            <div v-if="notifLog.length > 0" class="max-h-48 overflow-y-auto rounded-lg bg-(--ui-bg) border border-(--ui-border) p-2 space-y-1 text-[11px] font-mono leading-snug">
-              <div v-for="(entry, i) in notifLog" :key="i" class="break-all">
-                <span class="text-slate-600 shrink-0">{{ entry.ts.slice(11, 19) }}</span>
-                <span class="text-primary-400 ml-2">{{ entry.event }}</span>
-                <span v-if="entry.msg" class="text-(--ui-text-muted) ml-1"> {{ entry.msg }}</span>
-                <template v-for="(val, key) in entry.fields" :key="key">
-                  <span class="text-amber-400/70 ml-1">{{ key }}=</span><span class="text-(--ui-text-muted)">{{ val }}</span>
-                </template>
-              </div>
-            </div>
-            <p v-else class="text-xs text-slate-600 italic">No events yet. Schedule or test a notification to see activity here.</p>
-          </div>
-        </div>
-
-        <!-- Scheduled notifications -->
-        <div>
-          <div class="flex items-center justify-between px-4 py-3.5">
-            <div class="space-y-0.5">
-              <p class="text-sm font-medium">Scheduled notifications</p>
-              <p class="text-xs text-(--ui-text-dimmed)">
-                {{ scheduledNotifs.length > 0 ? `${scheduledNotifs.length} active` : 'None scheduled for today' }}
-              </p>
-            </div>
-            <UButton
-              size="sm" variant="ghost" color="neutral"
-              icon="i-heroicons-arrow-path"
-              :loading="pendingNotifsLoading"
-              aria-label="Refresh scheduled notifications"
-              @click="refreshPendingNotifs"
-            />
-          </div>
-          <ul v-if="scheduledNotifs.length > 0" class="border-t border-(--ui-border) divide-y divide-(--ui-border)/50 px-4 pb-3">
-            <li
-              v-for="(n, i) in scheduledNotifs"
-              :key="i"
-              class="flex items-center gap-3 py-2"
-            >
-              <UIcon name="i-heroicons-bell" class="w-3.5 h-3.5 shrink-0 text-(--ui-text-dimmed)" />
-              <span class="text-xs text-(--ui-text) min-w-0 truncate grow">{{ n.title }}</span>
-              <span class="text-xs font-mono text-(--ui-text-muted) shrink-0">{{ n.when }}</span>
-            </li>
-          </ul>
-        </div>
-
-        <!-- Storage estimate + OPFS files subsection -->
-        <div>
-          <div class="px-4 py-3.5 space-y-2">
-            <div class="flex items-center justify-between">
-              <div class="space-y-0.5">
-                <p class="text-sm font-medium">Storage</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Browser quota and current usage.</p>
-              </div>
-              <UButton size="sm" variant="ghost" color="neutral" icon="i-heroicons-arrow-path" @click="loadStorageEstimate" />
-            </div>
-            <template v-if="storageEstimate">
-              <div class="space-y-1">
-                <div class="flex justify-between text-xs text-(--ui-text-muted)">
-                  <span>{{ formatBytes(storageEstimate.usage ?? 0) }} used</span>
-                  <span>{{ formatBytes(storageEstimate.quota ?? 0) }} quota</span>
-                </div>
-                <div class="h-1.5 rounded-full bg-(--ui-bg-elevated) overflow-hidden">
-                  <div
-                    class="h-full rounded-full bg-primary-500 transition-all"
-                    :style="{ width: `${Math.min(100, ((storageEstimate.usage ?? 0) / (storageEstimate.quota ?? 1)) * 100).toFixed(1)}%` }"
-                  />
-                </div>
-              </div>
-              <div class="flex items-center justify-between gap-2">
-                <div class="flex items-center gap-1.5 text-xs">
-                  <UIcon
-                    :name="storagePersisted ? 'i-heroicons-lock-closed' : 'i-heroicons-lock-open'"
-                    class="w-3.5 h-3.5"
-                    :class="storagePersisted ? 'text-green-400' : 'text-amber-400'"
-                  />
-                  <span :class="storagePersisted ? 'text-green-400' : 'text-amber-400'">
-                    {{ storagePersisted === null ? 'Checking…' : storagePersisted ? 'Persistent storage' : 'Storage not persisted' }}
-                  </span>
-                </div>
-                <UButton
-                  v-if="storagePersisted === false"
-                  size="xs" variant="soft" color="primary"
-                  @click="requestPersistStorage"
-                >
-                  Request
-                </UButton>
-              </div>
-            </template>
-          </div>
-
-          <!-- OPFS files subsection -->
-          <div class="border-t border-(--ui-border)/60">
-            <button class="w-full flex items-center justify-between px-4 py-2.5 text-left" @click="toggleOpfs">
-              <p class="text-xs text-(--ui-text-dimmed)">OPFS files</p>
-              <div class="flex items-center gap-2 shrink-0">
-                <span v-if="opfsFiles.length > 0" class="text-xs text-(--ui-text-dimmed)">{{ opfsFiles.length }} file{{ opfsFiles.length !== 1 ? 's' : '' }}</span>
-                <UIcon :name="opfsOpen ? 'i-heroicons-chevron-down' : 'i-heroicons-chevron-right'" class="w-3.5 h-3.5 text-(--ui-text-dimmed)" />
-              </div>
-            </button>
-            <div v-if="opfsOpen" class="px-4 pb-3">
-              <div v-if="opfsLoading" class="flex items-center gap-2 text-xs text-(--ui-text-dimmed)">
-                <UIcon name="i-heroicons-arrow-path" class="w-3.5 h-3.5 animate-spin" /> Scanning…
-              </div>
-              <div v-else-if="opfsFiles.length === 0" class="text-xs text-(--ui-text-dimmed)">No files found.</div>
-              <ul v-else class="space-y-1.5">
-                <li v-for="f in opfsFiles" :key="f.path" class="flex items-center justify-between gap-4">
-                  <span class="text-[11px] font-mono text-(--ui-text-toned) truncate">{{ f.path }}</span>
-                  <span class="text-[11px] font-mono text-(--ui-text-dimmed) shrink-0">{{ formatBytes(f.size) }}</span>
-                </li>
-              </ul>
-              <button class="mt-3 text-xs text-slate-600 hover:text-(--ui-text-muted) flex items-center gap-1 transition-colors" @click="loadOpfsFiles">
-                <UIcon name="i-heroicons-arrow-path" class="w-3 h-3" /> Refresh
-              </button>
-            </div>
-          </div>
-        </div>
-
-        <!-- Remote debugging (APK) — Android native only -->
-        <div v-if="isAndroid" class="px-4 py-3.5 space-y-1">
-          <p class="text-sm font-medium">Remote debugging (APK)</p>
-          <ol class="text-xs text-(--ui-text-dimmed) list-decimal list-inside space-y-0.5">
-            <li>Enable <span class="text-(--ui-text-toned)">Developer Options</span> and <span class="text-(--ui-text-toned)">USB Debugging</span> on your Android device.</li>
-            <li>Connect via USB and open <span class="text-(--ui-text-toned) font-mono">chrome://inspect</span> in desktop Chrome.</li>
-            <li>Find the Habitat WebView under <span class="text-(--ui-text-toned)">Remote Target</span> and click <span class="text-(--ui-text-toned)">inspect</span>.</li>
-          </ol>
-          <p class="text-[11px] text-slate-600 mt-1">Console logs from this page will appear in the DevTools console.</p>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Here be dragons (collapsible) ─────────────────────────────────────── -->
-    <section class="space-y-2">
-      <button class="w-full flex items-center justify-between px-1 py-0.5" @click="dragonsOpen = !dragonsOpen">
-        <p class="text-xs font-semibold uppercase tracking-wider text-red-900/70">🐉 Here be dragons</p>
-        <UIcon :name="dragonsOpen ? 'i-heroicons-chevron-up' : 'i-heroicons-chevron-down'" class="w-3.5 h-3.5 text-red-900/50" />
-      </button>
-      <UCard v-if="dragonsOpen" :ui="{ root: 'rounded-2xl ring-1 ring-red-900/30', body: 'p-0 sm:p-0 divide-y divide-slate-800' }">
-
-        <!-- Strict network isolation -->
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5 mr-4">
-            <p class="text-sm font-medium">Strict network isolation</p>
-            <p class="text-xs text-(--ui-text-dimmed)">
-              Blocks all fetch and XHR. Allows cached assets and on-device SQLite.
-              <template v-if="strictCspSessionActive">
-                <span class="text-orange-400">Active this session — reload to change.</span>
-              </template>
-              <template v-else>
-                Requires a reload to activate or deactivate.
-              </template>
-            </p>
-          </div>
-          <div class="flex items-center gap-2 shrink-0">
-            <template v-if="strictCspSessionActive">
-              <UBadge color="warning" variant="subtle" size="xs" class="shrink-0">Active</UBadge>
-              <UButton
-                size="sm" variant="ghost" color="neutral"
-                icon="i-heroicons-arrow-path"
-                aria-label="Disable strict CSP and reload"
-                @click="disableStrictCspAndReload"
-              />
-            </template>
-            <USwitch
-              v-else
-              :model-value="appSettings.strictCsp"
-              aria-label="Enable strict network isolation"
-              @update:model-value="enableStrictCspAndReload"
-            />
-          </div>
-        </div>
-
-        <!-- Force reload — PWA only -->
-        <div v-if="!isNativeApp" class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium">Force reload</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Unregister service worker, clear JS/CSS caches, and reload. OPFS data is preserved.</p>
-          </div>
-          <UButton
-            size="sm" variant="ghost" color="neutral"
-            icon="i-heroicons-arrow-path" :loading="forceReloading" class="shrink-0"
-            @click="forceReload"
-          />
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium text-red-400">Clear app data</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Selectively delete habits, check-ins, scribbles, or voice notes.</p>
-          </div>
-          <span
-            class="shrink-0"
-            :class="!db.isAvailable ? 'cursor-not-allowed' : ''"
-            :title="!db.isAvailable ? 'Database not ready' : undefined"
-          >
-            <UButton
-              icon="i-heroicons-trash"
-              variant="ghost" color="error" size="sm"
-              :disabled="!db.isAvailable"
-              @click="showClearModal = true"
-            />
-          </span>
-        </div>
-
-        <div class="flex items-center justify-between px-4 py-3.5">
-          <div class="space-y-0.5">
-            <p class="text-sm font-medium text-red-400">Clear OPFS storage</p>
-            <p class="text-xs text-(--ui-text-dimmed)">Wipe all on-device file system storage including the database.</p>
-          </div>
-          <span
-            class="shrink-0"
-            :class="!db.isAvailable ? 'cursor-not-allowed' : ''"
-            :title="!db.isAvailable ? 'Database not ready' : undefined"
-          >
-            <UButton
-              icon="i-heroicons-fire"
-              variant="ghost" color="error" size="sm"
-              :disabled="!db.isAvailable"
-              @click="showNukeModal = true"
-            />
-          </span>
-        </div>
-
-      </UCard>
-    </section>
-
-    <!-- ── Philosophy note ───────────────────────────────────────────────────── -->
-    <p class="text-xs text-(--ui-text-dimmed) text-center leading-relaxed px-4 pb-2">
-      Habitat is built around the idea that habits should fit your life — not the other way around.
-      There are no streaks to protect, no daily guilt. Just a quiet space to notice what you do,
-      and gently do more of what matters.
-    </p>
-
-    <!-- ── Open source licenses modal ───────────────────────────────────────── -->
-    <UModal v-model:open="showLicensesModal">
-      <template #content>
-        <div class="p-5 space-y-4 flex flex-col max-h-[80vh]">
-          <div class="flex items-center justify-between shrink-0">
-            <h3 class="font-semibold text-(--ui-text)">Open source licenses</h3>
-            <UButton icon="i-heroicons-x-mark" variant="ghost" color="neutral" size="sm" @click="showLicensesModal = false" />
-          </div>
-          <div v-if="licensesLoading" class="flex items-center justify-center py-8 text-(--ui-text-dimmed) text-sm">
-            Loading…
-          </div>
-          <div v-else-if="licensesError" class="text-sm text-red-400 py-4">
-            Could not load licenses. Only available in production builds.
-          </div>
-          <ul v-else class="overflow-y-auto divide-y divide-slate-800 -mx-5 px-5">
-            <li v-for="pkg in licenses" :key="pkg.name" class="py-2.5 flex items-baseline justify-between gap-3">
-              <div class="min-w-0">
-                <a
-                  v-if="pkg.homepage"
-                  :href="pkg.homepage"
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  class="text-sm text-(--ui-text) hover:text-primary-400 transition-colors truncate block"
-                >{{ pkg.name }}</a>
-                <span v-else class="text-sm text-(--ui-text) truncate block">{{ pkg.name }}</span>
-                <span class="text-xs text-(--ui-text-dimmed) font-mono">v{{ pkg.version }}</span>
-              </div>
-              <UBadge
-                :label="pkg.license ?? 'Unknown'"
-                variant="subtle"
-                :color="pkg.license === 'MIT' ? 'primary' : 'neutral'"
-                size="xs"
-                class="shrink-0 font-mono rounded-full"
-              />
-            </li>
-          </ul>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Install instructions modal ───────────────────────────────────────── -->
-    <UModal v-model:open="showInstallModal">
-      <template #content>
-        <div class="p-5 space-y-4">
-          <div class="flex items-center justify-between">
-            <h3 class="font-semibold text-(--ui-text)">Install Habitat</h3>
-            <UButton icon="i-heroicons-x-mark" variant="ghost" color="neutral" size="sm" @click="showInstallModal = false" />
-          </div>
-
-          <!-- iOS Safari -->
-          <template v-if="isIosSafari">
-            <ol class="space-y-3 text-sm text-(--ui-text-toned)">
-              <li class="flex items-start gap-3">
-                <span class="text-primary-400 font-semibold shrink-0">1.</span>
-                <span>Tap the <strong class="text-(--ui-text)">Share</strong> button
-                  <UIcon name="i-heroicons-arrow-up-on-square" class="inline w-4 h-4 align-text-bottom text-(--ui-text)" />
-                  in the Safari toolbar</span>
-              </li>
-              <li class="flex items-start gap-3">
-                <span class="text-primary-400 font-semibold shrink-0">2.</span>
-                <span>Scroll down and tap <strong class="text-(--ui-text)">Add to Home Screen</strong></span>
-              </li>
-              <li class="flex items-start gap-3">
-                <span class="text-primary-400 font-semibold shrink-0">3.</span>
-                <span>Tap <strong class="text-(--ui-text)">Add</strong> to confirm</span>
-              </li>
-            </ol>
-          </template>
-
-          <!-- Chromium but no deferred prompt (not yet offered, or previously dismissed) -->
-          <template v-else-if="isChromiumNoPrompt">
-            <p class="text-sm text-(--ui-text-muted)">Chrome hasn't offered an install prompt yet — use the browser menu instead:</p>
-            <ol class="space-y-3 text-sm text-(--ui-text-toned)">
-              <li class="flex items-start gap-3">
-                <span class="text-primary-400 font-semibold shrink-0">1.</span>
-                <span>Tap the <strong class="text-(--ui-text)">⋮ menu</strong> in the top-right corner of Chrome</span>
-              </li>
-              <li class="flex items-start gap-3">
-                <span class="text-primary-400 font-semibold shrink-0">2.</span>
-                <span>Choose <strong class="text-(--ui-text)">Install app</strong> or <strong class="text-(--ui-text)">Add to Home Screen</strong></span>
-              </li>
-            </ol>
-          </template>
-
-          <!-- Unsupported browser -->
-          <template v-else>
-            <p class="text-sm text-(--ui-text-muted)">
-              Your current browser doesn't support PWA installation.
-              Open Habitat in <strong class="text-(--ui-text)">Chrome</strong>, <strong class="text-(--ui-text)">Edge</strong>,
-              <strong class="text-(--ui-text)">Brave</strong>, or <strong class="text-(--ui-text)">Opera</strong>
-              to install it as a standalone app.
-            </p>
-            <p class="text-xs text-(--ui-text-dimmed)">
-              These browsers also provide OPFS storage and background notifications.
-            </p>
-          </template>
-
-          <div class="flex justify-end">
-            <UButton variant="ghost" color="neutral" @click="showInstallModal = false">Close</UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Export JSON modal ─────────────────────────────────────────────────── -->
-    <UModal v-model:open="showExportModal">
-      <template #content>
-        <div class="p-5 space-y-4">
-          <div class="flex items-center justify-between">
-            <h3 class="font-semibold text-(--ui-text)">Export data</h3>
-            <UButton icon="i-heroicons-x-mark" variant="ghost" color="neutral" size="sm" @click="showExportModal = false" />
-          </div>
-
-          <!-- Select all toggle -->
-          <label class="flex items-center gap-2.5 cursor-pointer">
-            <input
-              type="checkbox"
-              class="accent-primary-500"
-              :checked="exportAllSelected"
-              :indeterminate="!exportAllSelected && !exportNoneSelected"
-              @change="toggleExportAll"
-            />
-            <span class="text-sm text-(--ui-text-toned) font-medium">Select all</span>
-          </label>
-
-          <!-- Groups -->
-          <div class="space-y-3">
-            <div v-for="group in EXPORT_GROUPS" :key="group.label" class="space-y-1.5">
-              <p class="text-xs font-semibold uppercase tracking-wider text-(--ui-text-dimmed)">{{ group.label }}</p>
-              <div class="border border-(--ui-border) rounded-xl divide-y divide-slate-800 overflow-hidden">
-                <label
-                  v-for="item in group.items"
-                  :key="item.key"
-                  class="flex items-center gap-3 px-3.5 py-2.5 cursor-pointer hover:bg-(--ui-bg-elevated)/40 transition-colors"
-                >
-                  <input v-model="exportSel[item.key]" type="checkbox" class="accent-primary-500 shrink-0" />
-                  <span class="text-sm text-(--ui-text-toned)">{{ item.label }}</span>
-                </label>
-              </div>
-            </div>
-          </div>
-
-          <!-- FK errors -->
-          <ul v-if="exportErrors.length > 0" class="space-y-1">
-            <li
-              v-for="err in exportErrors"
-              :key="err"
-              class="text-xs text-red-400 flex items-center gap-1.5"
-            >
-              <UIcon name="i-heroicons-exclamation-circle" class="w-3.5 h-3.5 shrink-0" />
-              {{ err }}
-            </li>
-          </ul>
-
-          <div class="flex justify-end gap-2 pt-1">
-            <UButton variant="ghost" color="neutral" size="sm" @click="showExportModal = false">Cancel</UButton>
-            <UButton
-              size="sm"
-              icon="i-heroicons-arrow-down-tray"
-              :loading="exporting"
-              :disabled="exportNoneSelected"
-              @click="downloadJson"
-            >
-              Download
-            </UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Jots export modal ─────────────────────────────────────────────────── -->
-    <UModal v-model:open="showJotsExportModal">
-      <template #content>
-        <div class="p-5 space-y-4">
-          <div class="flex items-center justify-between">
-            <h3 class="font-semibold text-(--ui-text)">Export Jots</h3>
-            <UButton icon="i-heroicons-x-mark" variant="ghost" color="neutral" size="sm" @click="showJotsExportModal = false" />
-          </div>
-
-          <p class="text-sm text-(--ui-text-muted)">Choose which types to include. All selected types are bundled into a single <span class="font-mono">.zip</span>.</p>
-
-          <div class="space-y-3">
-            <label class="flex items-start gap-3 cursor-pointer">
-              <input v-model="jotsExportSel.text" type="checkbox" class="mt-0.5 accent-primary-500" />
-              <div>
-                <p class="text-sm font-medium text-(--ui-text)">Text notes</p>
-                <p class="text-xs text-(--ui-text-dimmed)">One <span class="font-mono">.txt</span> per note in <span class="font-mono">text/</span> + <span class="font-mono">jots.json</span></p>
-              </div>
-            </label>
-            <label class="flex items-start gap-3 cursor-pointer">
-              <input v-model="jotsExportSel.voice" type="checkbox" class="mt-0.5 accent-primary-500" />
-              <div>
-                <p class="text-sm font-medium text-(--ui-text)">Voice recordings</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Audio files in a <span class="font-mono">voice/</span> folder</p>
-              </div>
-            </label>
-            <label class="flex items-start gap-3 cursor-pointer">
-              <input v-model="jotsExportSel.images" type="checkbox" class="mt-0.5 accent-primary-500" />
-              <div>
-                <p class="text-sm font-medium text-(--ui-text)">Images</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Photos in an <span class="font-mono">images/</span> folder</p>
-              </div>
-            </label>
-          </div>
-
-          <div class="flex gap-2 pt-1">
-            <UButton
-              class="flex-1"
-              icon="i-heroicons-arrow-down-tray"
-              :loading="exportingJots"
-              :disabled="!jotsExportSel.text && !jotsExportSel.voice && !jotsExportSel.images"
-              @click="exportJotsZip"
-            >
-              Download .zip
-            </UButton>
-            <UButton variant="outline" color="neutral" @click="showJotsExportModal = false">Cancel</UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Import JSON modal ──────────────────────────────────────────────────── -->
-    <UModal v-model:open="showImportModal">
-      <template #content>
-        <!-- Error state -->
-        <div v-if="importError" class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-red-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-exclamation-circle" class="w-5 h-5 text-red-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">Cannot import</p>
-              <p class="text-sm text-(--ui-text-muted)">{{ importError }}</p>
-            </div>
-          </div>
-          <div class="flex justify-end">
-            <UButton variant="ghost" color="neutral" @click="showImportModal = false">Close</UButton>
-          </div>
-        </div>
-
-        <!-- Success state -->
-        <div v-else-if="importDone" class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-green-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-check-circle" class="w-5 h-5 text-green-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">Import complete</p>
-              <p class="text-sm text-(--ui-text-muted)">Records merged. Reload to see all updates.</p>
-            </div>
-          </div>
-          <div class="flex justify-end">
-            <UButton @click="() => { showImportModal = false; reloadPage() }">Reload</UButton>
-          </div>
-        </div>
-
-        <!-- Preview / confirm state -->
-        <div v-else-if="importPreview" class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-primary-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-arrow-up-tray" class="w-5 h-5 text-primary-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">Import data?</p>
-              <p class="text-sm text-(--ui-text-muted)">New records will be added; existing IDs are skipped.</p>
-            </div>
-          </div>
-
-          <!-- Summary -->
-          <div class="border border-(--ui-border) rounded-xl p-3 space-y-1">
-            <p class="text-xs text-(--ui-text-dimmed) mb-2">
-              Version {{ importPreview.version }} export from {{ new Date(importPreview.exported_at).toLocaleDateString() }}
-            </p>
-            <template v-for="group in EXPORT_GROUPS" :key="group.label">
-              <template v-for="item in group.items" :key="item.key">
-                <div
-                  v-if="(importPreview[item.key] ?? []).length > 0"
-                  class="flex items-center justify-between text-xs"
-                >
-                  <span class="text-(--ui-text-muted)">{{ item.label }}</span>
-                  <span class="font-mono text-(--ui-text-toned)">{{ (importPreview[item.key] ?? []).length }}</span>
-                </div>
-              </template>
-            </template>
-          </div>
-
-          <div class="flex justify-end gap-2 pt-1">
-            <UButton variant="ghost" color="neutral" @click="showImportModal = false">Cancel</UButton>
-            <UButton :loading="importing" @click="confirmImport">Import</UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Confirm clear app data ────────────────────────────────────────────── -->
-    <UModal v-model:open="showClearModal">
-      <template #content>
-        <div class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-red-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-exclamation-triangle" class="w-5 h-5 text-red-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">Clear app data?</p>
-              <p class="text-sm text-(--ui-text-muted)">Select what to delete. This cannot be undone.</p>
-            </div>
-          </div>
-
-          <!-- Checkboxes -->
-          <div class="space-y-1 border border-(--ui-border) rounded-xl divide-y divide-slate-800 overflow-hidden">
-            <label
-              v-for="item in clearItems"
-              :key="item.key"
-              class="flex items-start gap-3 px-3.5 py-3 cursor-pointer hover:bg-(--ui-bg-elevated)/40 transition-colors"
-            >
-              <input
-                v-model="clearSelection[item.key]"
-                type="checkbox"
-                class="mt-0.5 accent-red-500 shrink-0"
-              />
-              <div class="min-w-0">
-                <p class="text-sm font-medium text-(--ui-text) leading-snug">{{ item.label }}</p>
-                <p class="text-xs text-(--ui-text-dimmed) leading-snug">{{ item.description }}</p>
-              </div>
-            </label>
-          </div>
-
-          <div class="flex justify-end gap-2 pt-1">
-            <UButton variant="ghost" color="neutral" @click="showClearModal = false">Cancel</UButton>
-            <UButton
-              color="error"
-              :loading="clearing"
-              :disabled="nothingSelected"
-              @click="clearAppData"
-            >
-              Delete selected
-            </UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Confirm nuke OPFS ───────────────────────────────────────────────────── -->
-    <UModal v-model:open="showNukeModal">
-      <template #content>
-        <!-- Done state (wipe-only path) -->
-        <div v-if="wiped" class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-green-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-check-circle" class="w-5 h-5 text-green-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">All data wiped</p>
-              <p class="text-sm text-(--ui-text-muted)">Storage has been cleared. You can safely close this tab.</p>
-            </div>
-          </div>
-        </div>
-
-        <!-- Confirm state -->
-        <div v-else class="p-5 space-y-4">
-          <div class="flex items-start gap-3">
-            <div class="w-10 h-10 rounded-full bg-red-500/10 flex items-center justify-center flex-shrink-0">
-              <UIcon name="i-heroicons-fire" class="w-5 h-5 text-red-400" />
-            </div>
-            <div class="space-y-1">
-              <p class="font-semibold">Wipe OPFS storage?</p>
-              <p class="text-sm text-(--ui-text-muted)">
-                Removes every file in the browser's origin private file system —
-                including the SQLite database. Voice recordings and check-in entries
-                will also be cleared.
-              </p>
-            </div>
-          </div>
-          <div class="flex justify-end gap-2 pt-1">
-            <UButton variant="ghost" color="neutral" @click="showNukeModal = false">Cancel</UButton>
-            <UButton color="error" variant="ghost" :loading="nuking" @click="nukeOpfs(false)">Wipe only</UButton>
-            <UButton color="error" :loading="nuking" @click="nukeOpfs(true)">Wipe &amp; reload</UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
-
-    <!-- ── Health setup modal ───────────────────────────────────────────────── -->
-    <UModal v-model:open="showHealthSetup">
-      <template #content>
-        <div class="p-5 space-y-5">
-          <div>
-            <h3 class="text-lg font-semibold">Set up Health Tracking</h3>
-            <p class="text-sm text-(--ui-text-muted) mt-0.5">Choose what you'd like to track. Habits are created in your habit list.</p>
-          </div>
-
-          <!-- Steps -->
-          <div class="space-y-3">
-            <label class="flex items-center gap-3 cursor-pointer">
-              <input v-model="healthSetup.enableSteps" type="checkbox" class="rounded" />
-              <div>
-                <p class="text-sm font-medium">Track Steps</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Creates a daily NUMERIC habit with your step goal.</p>
-              </div>
-            </label>
-            <div v-if="healthSetup.enableSteps" class="ml-7">
-              <UFormField label="Daily step goal">
-                <UInput
-                  v-model.number="healthSetup.stepGoal"
-                  type="number"
-                  min="1000"
-                  step="500"
-                  class="w-40"
-                />
-              </UFormField>
-            </div>
-          </div>
-
-          <!-- Meals -->
-          <div class="space-y-3">
-            <label class="flex items-center gap-3 cursor-pointer">
-              <input v-model="healthSetup.enableMeals" type="checkbox" class="rounded" />
-              <div>
-                <p class="text-sm font-medium">Track Meals</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Creates LIMIT habits for calorie tracking per meal.</p>
-              </div>
-            </label>
-            <div v-if="healthSetup.enableMeals" class="ml-7 space-y-2">
-              <div v-for="(meal, i) in healthSetup.meals" :key="i" class="flex items-center gap-2">
-                <UInput v-model="meal.name" placeholder="Meal name" class="flex-1" />
-                <UInput
-                  v-model.number="meal.calories"
-                  type="number"
-                  min="0"
-                  step="50"
-                  class="w-24"
-                />
-                <span class="text-xs text-(--ui-text-dimmed) shrink-0">kcal</span>
-                <button class="text-slate-700 hover:text-red-400 transition-colors shrink-0" @click="removeMeal(i)">
-                  <UIcon name="i-heroicons-x-mark" class="w-4 h-4" />
-                </button>
-              </div>
-              <button class="text-xs text-(--ui-text-dimmed) hover:text-(--ui-text-muted) flex items-center gap-1 mt-1" @click="addMeal">
-                <UIcon name="i-heroicons-plus" class="w-3 h-3" /> Add meal
-              </button>
-            </div>
-          </div>
-
-          <!-- Water -->
-          <div class="space-y-3">
-            <label class="flex items-center gap-3 cursor-pointer">
-              <input v-model="healthSetup.enableWater" type="checkbox" class="rounded" />
-              <div>
-                <p class="text-sm font-medium">Track Water</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Creates a daily NUMERIC habit to count glasses.</p>
-              </div>
-            </label>
-            <div v-if="healthSetup.enableWater" class="ml-7">
-              <UFormField label="Daily glasses goal">
-                <UInput
-                  v-model.number="healthSetup.waterGoal"
-                  type="number"
-                  min="1"
-                  max="20"
-                  step="1"
-                  class="w-32"
-                />
-              </UFormField>
-            </div>
-          </div>
-
-          <!-- Sleep -->
-          <div class="space-y-3">
-            <label class="flex items-center gap-3 cursor-pointer">
-              <input v-model="healthSetup.enableSleep" type="checkbox" class="rounded" />
-              <div>
-                <p class="text-sm font-medium">Track Sleep</p>
-                <p class="text-xs text-(--ui-text-dimmed)">Creates a daily NUMERIC habit for hours slept.</p>
-              </div>
-            </label>
-            <div v-if="healthSetup.enableSleep" class="ml-7">
-              <UFormField label="Sleep goal (hours)">
-                <UInput
-                  v-model.number="healthSetup.sleepGoal"
-                  type="number"
-                  min="1"
-                  max="12"
-                  step="0.5"
-                  class="w-32"
-                />
-              </UFormField>
-            </div>
-          </div>
-
-          <div class="flex justify-end gap-2 pt-1">
-            <UButton variant="ghost" color="neutral" @click="showHealthSetup = false">Skip</UButton>
-            <UButton
-              color="primary"
-              :disabled="!healthSetup.enableSteps && !healthSetup.enableMeals && !healthSetup.enableWater && !healthSetup.enableSleep"
-              :loading="creatingHealth"
-              @click="confirmHealthSetup"
-            >
-              Create Habits
-            </UButton>
-          </div>
-        </div>
-      </template>
-    </UModal>
+  <div class="sm:flex sm:gap-6">
+
+    <!-- Sidebar: desktop only -->
+    <aside class="hidden sm:flex sm:flex-col sm:w-44 sm:shrink-0">
+      <h2 class="text-2xl font-bold px-3 pb-3">Settings</h2>
+      <nav class="sticky top-4 flex flex-col gap-0.5">
+        <NuxtLink
+          v-for="s in SETTINGS_SECTIONS"
+          :key="s.to"
+          :to="s.to"
+          class="flex items-center gap-2 px-3 py-2 rounded-xl text-sm font-medium transition-colors"
+          :class="route.path.startsWith(s.to)
+            ? 'bg-(--ui-bg-elevated) text-(--ui-text)'
+            : 'text-(--ui-text-muted) hover:text-(--ui-text-toned) hover:bg-(--ui-bg-muted)'"
+        >
+          <UIcon :name="s.icon" class="w-4 h-4 shrink-0" />
+          {{ s.label }}
+        </NuxtLink>
+        <p class="mt-4 px-3 text-[11px] text-(--ui-text-dimmed) leading-relaxed">
+          Habits should fit your life — not the other way around. No streaks, no guilt. Just a quiet space to notice what matters.
+        </p>
+      </nav>
+    </aside>
+
+    <!-- Content -->
+    <div class="flex-1 min-w-0">
+      <NuxtPage />
+    </div>
 
   </div>
 </template>
